~10% performance boost by raising tile size to 8x8

+++ src/main/java/net/daporkchop/bedrock/mode/BedrockFullScanner.java
@@ -40,8 +40,8 @@ public class BedrockFullScanner implements TileScanner {
     }
 
     @Override
-    public int scan(int tileX, int tileZ) {
-        int bits = 0;
+    public long scan(int tileX, int tileZ) {
+        long bits = 0L;
         final byte[][] patterns = this.patterns;
         final int numPatterns = patterns.length;
 
@@ -80,7 +80,7 @@ public class BedrockFullScanner implements TileScanner {
                         state = updateBedrock(state);
                     }
 
-                    bits |= 1 << ((subX << TILE_SHIFT) | subZ);
+                    bits |= 1L << ((subX << TILE_SHIFT) | subZ);
                     break;
                 }
             }
