+++ src/main/java/net/daporkchop/bedrock/mode/BedrockSubScanner.java
@@ -42,8 +42,8 @@ public class BedrockSubScanner implements TileScanner {
     }
 
     @Override
-    public int scan(int tileX, int tileZ) {
-        int bits = 0;
+    public long scan(int tileX, int tileZ) {
+        long bits = 0L;
         final byte[][] patterns = this.patterns;
         final int numPatterns = patterns.length;
 
@@ -75,7 +75,7 @@ public class BedrockSubScanner implements TileScanner {
 
                             if (!wrong) {
                                 //if we've gotten this far, a match has been found
-                                bits |= 1 << ((subX << TILE_SHIFT) | subZ);
+                                bits |= 1L << ((subX << TILE_SHIFT) | subZ);
                                 break PATTERN;
                             }
                         }
